server/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-filter==23.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2==2.9.5
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
shortuuid==1.0.11
sqlparse==0.4.3
